+++ test/change_health/response/claim/report/report_835_service_line_test.rb
@@ -15,33 +15,9 @@ class Report835ServiceLineTest < Minitest::Test
          )
       end
 
-      let(:health_care_check_remark_code) do
-         ChangeHealth::Response::Claim::Report835HealthCareCheckRemarkCode.new(
-         remark_code: "M1",
-         code_list_qualifier_code_value: "Claim Payment Remark Codes",
-         code_list_qualifier_code: "HI"
-         )
-      end
-
-      let(:health_care_check_remark_code_2) do
-         ChangeHealth::Response::Claim::Report835HealthCareCheckRemarkCode.new(
-         remark_code: "N510"
-         )
-      end
-
       let(:claim_information) do
          ChangeHealth::Response::Claim::Report835ServiceLine.new(adjudicated_procedure_code: 'J1745', allowed_actual: 30_720.0, line_item_charge_amount: 48_000.0, line_item_provider_payment_amount: '18432', service_adjustments: [service_adjustments, service_adjustments_2],
          health_care_check_remark_codes: [])
-      end
-
-      let(:claim_information_with_remark_codes) do
-         ChangeHealth::Response::Claim::Report835ServiceLine.new(adjudicated_procedure_code: 'J1745', allowed_actual: 30_720.0, line_item_charge_amount: 48_000.0, line_item_provider_payment_amount: '18432', service_adjustments: [service_adjustments, service_adjustments_2],
-         health_care_check_remark_codes: [health_care_check_remark_code, health_care_check_remark_code_2])
-      end
-
-      let(:claim_information_with_nil_codes) do
-         ChangeHealth::Response::Claim::Report835ServiceLine.new(adjudicated_procedure_code: 'J1745', allowed_actual: 30_720.0, line_item_charge_amount: 48_000.0, line_item_provider_payment_amount: '18432', service_adjustments: [service_adjustments, service_adjustments_2],
-         health_care_check_remark_codes: nil)
       end
 
       it 'creates adjustments correctly when there are multiple adjustments in a group code' do
@@ -75,90 +51,6 @@ class Report835ServiceLineTest < Minitest::Test
 
          actual_result = claim_information.create_adjustment_detail_array
          assert_equal(expected_answer, actual_result)
-      end
-
-      it 'creates adjustments correctly when there are remark codes' do
-         expected_result = [
-            {
-               adjustmentDetails: [
-               {
-                  adjustmentReasonCode: "45",
-                  adjustmentAmount: "180.82"
-               },
-               {
-                  adjustmentReasonCode: "253",
-                  adjustmentAmount: "13.24"
-               },
-               {
-                  adjustmentReasonCode: "59", adjustmentAmount: "827.59"
-               }
-               ],
-               adjustmentGroupCode: "PR"
-            },
-            {
-               adjustmentDetails: [
-               {
-                  adjustmentReasonCode: "2",
-                  adjustmentAmount: "165.52"
-               }
-               ],
-               adjustmentGroupCode: "CO"
-            },
-            {
-               adjustmentDetails: [
-               {
-                  adjustmentReasonCode: "M1",
-                  adjustmentAmount: "0"
-               }
-               ],
-               adjustmentGroupCode: ""
-            },
-            {
-               adjustmentDetails: [
-               {
-                  adjustmentReasonCode: "N510",
-                  adjustmentAmount: "0"
-               }
-               ],
-               adjustmentGroupCode: ""
-            }
-         ]
-
-         actual_result = claim_information_with_remark_codes.create_adjustment_detail_array
-         assert_equal(expected_result, actual_result)
-      end
-
-      it 'creates adjustments correctly when there are remark codes' do
-         expected_result = [
-            {
-               adjustmentDetails: [
-               {
-                  adjustmentReasonCode: "45",
-                  adjustmentAmount: "180.82"
-               },
-               {
-                  adjustmentReasonCode: "253",
-                  adjustmentAmount: "13.24"
-               },
-               {
-                  adjustmentReasonCode: "59", adjustmentAmount: "827.59"
-               }
-               ],
-               adjustmentGroupCode: "PR"
-            },
-            {
-               adjustmentDetails: [
-               {
-                  adjustmentReasonCode: "2",
-                  adjustmentAmount: "165.52"
-               }
-               ],
-               adjustmentGroupCode: "CO"
-            }
-         ]
-
-         actual_result = claim_information_with_nil_codes.create_adjustment_detail_array
-         assert_equal(expected_result, actual_result)
       end
    end
 end